requirements.txt
PyQt5>=5.15.0
Pillow>=8.0.0
reportlab>=3.5.0
pdfkit>=1.0.0
